+++ line-oauth/package.js
@@ -2,7 +2,7 @@
 Package.describe({
   name: 'storyteller:line-oauth',
   summary: 'LINE OAuth flow',
-  version: '1.4.1-alpha.1',
+  version: '1.4.1-alpha.2',
   git: 'https://github.com/StorytellerCZ/meteor-accounts-line'
 });
 
@@ -11,8 +11,8 @@ Npm.depends({
 });
 
 Package.onUse(api => {
-  api.versionsFrom(['2.9.0', '3.0-alpha.11']);
-  api.use('zodern:types@1.0.9', 'server')
+  api.versionsFrom(['2.9.0', '3.0-rc.0']);
+  api.use('zodern:types@1.0.13', 'server')
   api.use(['ecmascript', 'typescript'], ['client', 'server']);
   api.use('oauth2', ['client', 'server']);
   api.use('oauth', ['client', 'server']);