+++ line-oauth/CHANGELOG.md
@@ -1,4 +1,11 @@
 # Changelog
+## 1.4.2-alpha.3 - 2024-05-04
+* Meteor build targets added for 3.0-rc.0
+* Updated `zodern:types` to version 1.0.13
+
+## 1.4.1-alpha.2 - 2023-08-24
+* Use async call to `ServiceConfiguration`
+
 ## 1.4.1-alpha.1 - 2023-08-24
 * Added target of Meteor `3.0-alpha.11` to allow building in Meteor 3.0 alpha